# [FLINK-31843][runtime] Improve performance of SlotSelectionStrategy by bulk select.