--- a/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/AllocatedSlotPool.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/AllocatedSlotPool.java
@@ -133,7 +133,7 @@ public interface AllocatedSlotPool {
 
     /** Information about a free slot. */
     interface FreeSlotInfo {
-        SlotInfoWithUtilization asSlotInfo();
+        SlotInfo asSlotInfo();
 
         /**
          * Returns since when this slot is free.