--- a/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DeclarativeSlotPool.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DeclarativeSlotPool.java
@@ -113,7 +113,7 @@ public interface DeclarativeSlotPool {
      *
      * @return collection of free slot information
      */
-    Collection<SlotInfoWithUtilization> getFreeSlotsInformation();
+    Collection<SlotInfo> getFreeSlotsInformation();
 
     /**
      * Returns the free slot tracker.