--- a/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DeclarativeSlotPoolBridge.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DeclarativeSlotPoolBridge.java
@@ -440,6 +440,13 @@ public class DeclarativeSlotPoolBridge extends DeclarativeSlotPoolService implem
     }
 
     @Override
+    public FreeSlotInfoTracker getFreeSlotInfoTracker() {
+        assertRunningInMainThread();
+
+        return getDeclarativeSlotPool().getFreeSlotInfoTracker();
+    }
+
+    @Override
     public void disableBatchSlotRequestTimeoutCheck() {
         isBatchSlotRequestTimeoutCheckDisabled = true;
     }