--- a/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DeclarativeSlotPoolBridge.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DeclarativeSlotPoolBridge.java
@@ -423,7 +423,7 @@ public class DeclarativeSlotPoolBridge extends DeclarativeSlotPoolService implem
                 getDeclarativeSlotPool().getAllSlotsInformation();
         final Set<AllocationID> freeSlots =
                 getDeclarativeSlotPool().getFreeSlotsInformation().stream()
-                        .map(SlotInfoWithUtilization::getAllocationId)
+                        .map(SlotInfo::getAllocationId)
                         .collect(Collectors.toSet());
 
         return allSlotsInformation.stream()
@@ -433,7 +433,7 @@ public class DeclarativeSlotPoolBridge extends DeclarativeSlotPoolService implem
 
     @Override
     @Nonnull
-    public Collection<SlotInfoWithUtilization> getAvailableSlotsInformation() {
+    public Collection<SlotInfo> getAvailableSlotsInformation() {
         assertRunningInMainThread();
 
         return getDeclarativeSlotPool().getFreeSlotsInformation();