--- a/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DeclarativeSlotPoolService.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DeclarativeSlotPoolService.java
@@ -247,7 +247,7 @@ public class DeclarativeSlotPoolService implements SlotPoolService {
                                             slotInfo.getTaskManagerLocation()
                                                     .getResourceID()
                                                     .equals(taskManagerId))
-                            .map(SlotInfoWithUtilization::getAllocationId)
+                            .map(SlotInfo::getAllocationId)
                             .collect(Collectors.toSet());
 
             for (AllocationID allocationId : freeSlots) {