--- a/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DefaultDeclarativeSlotPool.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/slotpool/DefaultDeclarativeSlotPool.java
@@ -52,7 +52,6 @@ import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 
 /**
  * Default {@link DeclarativeSlotPool} implementation.
@@ -487,7 +486,7 @@ public class DefaultDeclarativeSlotPool implements DeclarativeSlotPool {
     @Override
     public void releaseIdleSlots(long currentTimeMillis) {
         final Collection<AllocatedSlotPool.FreeSlotInfo> freeSlotsInformation =
-                slotPool.getFreeSlotsInformation();
+                slotPool.getFreeSlotInfoTracker().getFreeSlotsWithIdleSinceInformation();
 
         ResourceCounter excessResources =
                 fulfilledResourceRequirements.subtract(totalResourceRequirements);
@@ -561,13 +560,6 @@ public class DefaultDeclarativeSlotPool implements DeclarativeSlotPool {
                         }
                     });
         }
-    }
-
-    @Override
-    public Collection<SlotInfo> getFreeSlotsInformation() {
-        return slotPool.getFreeSlotsInformation().stream()
-                .map(AllocatedSlotPool.FreeSlotInfo::asSlotInfo)
-                .collect(Collectors.toList());
     }
 
     @Override