## What is the purpose of the change
Improving the performance of the SlotSelectionStrategy by bulk selecting tasks in the same region.


## Brief change log
  - *Introduce FreeSlotCandidatesManager to manage available slots for candidates*
  - *Migerate SlotInfoWithUtilization to FreeSlotCandidatesManager for SlotSelectionStrategy*
  - *PhysicalSlotProvider support bulk select available slots*


## Verifying this change
  - *Added unit tests for FreeSlotCandidatesManager*
  - *Manually verified th change by wordcount job with 20000 parallelism*
  - *Compare the benchmark of handleGlobalFailureAndRestartAllTasks*

The benchmark result of handleGlobalFailureAndRestartAllTasks are:
master branch:
```
Benchmark                              (jobConfiguration)  Mode  Cnt      Score     Error  Units
handleGlobalFailureAndRestartAllTasks               BATCH    ss   20   2625.949 ? 204.485  ms/op
handleGlobalFailureAndRestartAllTasks           STREAMING    ss   20   8417.172 ? 543.077  ms/op
handleGlobalFailureAndRestartAllTasks        BATCH_EVENLY    ss   20   4914.222 ? 375.593  ms/op
handleGlobalFailureAndRestartAllTasks    STREAMING_EVENLY    ss   20  20335.352 ? 894.569  ms/op
```
after this pr 
```
Benchmark                              (jobConfiguration)  Mode  Cnt      Score      Error  Units
handleGlobalFailureAndRestartAllTasks               BATCH    ss   20   2096.568 ?   95.519  ms/op
handleGlobalFailureAndRestartAllTasks           STREAMING    ss   20   1727.778 ?  171.373  ms/op
handleGlobalFailureAndRestartAllTasks        BATCH_EVENLY    ss   20   5117.082 ?  341.998  ms/op
handleGlobalFailureAndRestartAllTasks    STREAMING_EVENLY    ss   20  17622.214 ? 1218.075  ms/op
```
The time cost of handleGlobalFailureAndRestartAllTasks for the STREAMING scene is reduced by 80%, and the BATCH and STREAMING_EVENLY scene is reduced by 20%.

However, BATCH_EVENLY has a slight performance regression (less than 5%). I think this is acceptable. Because: 1) Batch Evenly is a strategy that is unlikely to be used in production, batch tasks run for a short time, and resources can be released when they are finished; 2) This only affects the failover process, and part of batch tasks (with block shuffling) will not trigger the global Failover;

The entire benchmark results run in my local ECS are:
- master branch default benchmarks:  [master-default-jmh-result.csv](https://github.com/apache/flink/files/12136552/master-default-jmh-result.csv)
- master branch scheduler benchmarks: [master-scheduler-jmh-result.csv](https://github.com/apache/flink/files/12136553/master-scheduler-jmh-result.csv)
- changed branch default benchmarks: [slotchange-default-jmh-result.csv](https://github.com/apache/flink/files/12136554/slotchange-default-jmh-result.csv)
- changed branch scheduler benchmarks: [slotchange-scheduler-jmh-result.csv](https://github.com/apache/flink/files/12136555/slotchange-scheduler-jmh-result.csv)
- The results of org.apache.flink.benchmark.KeyByBenchmarks.tupleKeyBy for this report are very different. But it's fine after re-run it multiple times.


## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): (no)
  - The public API, i.e., is any changed class annotated with `@Public(Evolving)`:  (no)
  - The serializers:  (no)
  - The runtime per-record code paths (performance sensitive):  (no)
  - Anything that affects deployment or recovery: JobManager (and its components), Checkpointing, Kubernetes/Yarn, ZooKeeper: (no)
  - The S3 file system connector: (no)

## Documentation

  - Does this pull request introduce a new feature?  (no)