@huwh 2023-07-21
@KarmaGYZ Thanks for review, I have updated this pr, PTAL

Because the class name has changed, it's hard to create fixup commits without conflict, so I change the origin commit and force pushed. The main change is here 
 https://github.com/apache/flink/compare/8d753b63d4984cba13641b7a4652df3c1687c51a..67e21c99d13f6e9277bd39d2a91f7ea5abbd5a0d

After this chan, I rebase it into the master branch.