@KarmaGYZ 2023-07-21
@flinkbot run azure